Refactored Dashboard greeting component
